+++ packages/snippets/src/ant-design-x/index.ts
@@ -1,4 +1,7 @@
 export * from '@mixte/snippets/ant-design-x/bubble';
 export * from '@mixte/snippets/ant-design-x/sender';
+export {
+  useReactFunction,
+} from '@mixte/snippets/ant-design-x/utils';
 export * from '@mixte/snippets/ant-design-x/welcome';
 export * from '@mixte/snippets/ant-design-x/x-provider';