+++ packages/snippets/src/ant-design-x/src/utils.ts
@@ -2,7 +2,8 @@ import type { ReactElement } from 'react';
 import type { Component, VNodeChild } from 'vue';
 import { isFunction } from 'mixte';
 import React from 'react';
-import { applyPureVueInReact } from 'veaury';
+import { applyPureReactInVue, applyPureVueInReact } from 'veaury';
+import { shallowRef } from 'vue';
 
 /**
  * Vue 组件或插槽参数类型
@@ -20,4 +21,27 @@ export function renderVueCompOrSlot(comp?: VueCompOrSlot<any>): ReactElement {
   }
 
   return comp;
+}
+
+/**
+ * 在 Vue 环境中执行 React 函数并捕获其返回值
+ * 创建一个不渲染任何内容的React组件, 但可以获取函数执行的结果
+ * @param fn 要在React上下文中执行的函数
+ * @returns 包含不可见组件和状态引用的对象
+ */
+export function useReactFunction<
+  T,
+>(fn: () => T) {
+  const status = shallowRef<T>();
+
+  const Comp: Component = applyPureReactInVue(() => {
+    status.value = fn();
+
+    return null;
+  });
+
+  return {
+    Comp,
+    status,
+  };
 }